requirements.txt
beautifulsoup4
bitly-api-python==1.0.3
bs4
certifi
charset-normalizer
click
cloudscraper
flask
flask_restful
idna
itsdangerous
Jinja2
lk21
lxml
MarkupSafe
pyparsing
requests
requests-toolbelt
soupsieve
urllib3
waitress
Werkzeug